+++ native/agent/src/main/cpp/simplecallcounterprofiler.cpp
@@ -35,12 +35,12 @@ void SimpleCallCounterProfiler::methodEntry(int cnum, int mnum, jobject thread){
   auto methodId = Utils::getMethodId(cnum, mnum);
   auto info = getRuntime()->getCurrentThreadInfo();
   pthread_t threadKey = info.getProcessTid();
-  //unsigned long threadKey = (unsigned long)(*(long *)thread);
+  //unsigned long long threadKey = (unsigned long long)(*(long *)thread);
   auto it = statByThread.find(threadKey);
-  unordered_map<unsigned long, CallStatistics*> *stat = nullptr;
+  unordered_map<unsigned long long, CallStatistics*> *stat = nullptr;
   
   if(it==statByThread.end()){
-    stat = new unordered_map<unsigned long, CallStatistics*>();
+    stat = new unordered_map<unsigned long long, CallStatistics*>();
     statByThread.emplace(threadKey, stat);
   } else {
     stat = (it->second);
@@ -63,10 +63,10 @@ void SimpleCallCounterProfiler::methodEntry(int cnum, int mnum, jobject thread){
 void SimpleCallCounterProfiler::methodExit(int cnum, int mnum, jobject thread){
   auto info = getRuntime()->getCurrentThreadInfo();
   pthread_t threadKey = info.getProcessTid();
-  //unsigned long threadKey = (unsigned long)(*(long *)thread);
+  //unsigned long long threadKey = (unsigned long long)(*(long *)thread);
   auto it = statByThread.find(threadKey);
   
-  unordered_map<unsigned long, CallStatistics*> *stat = nullptr;
+  unordered_map<unsigned long long, CallStatistics*> *stat = nullptr;
   
   if(it!=statByThread.end()){
     stat = (it->second);
@@ -92,13 +92,13 @@ void SimpleCallCounterProfiler::printOnExit(){
   getRuntime()->agentGlobalLock();
   
   unsigned int calls1 = 0;
-  unsigned long total = 0;
+  unsigned long long total = 0;
 
   cout << "Threads " << statByThread.size() <<endl;
   
   for(auto it=statByThread.begin();it!=statByThread.end();it++){
       cout << "Thread " << it->first << " " << endl;
-      unordered_map<unsigned long, CallStatistics*> *calls = it->second;
+      unordered_map<unsigned long long, CallStatistics*> *calls = it->second;
       cout << "\t" << "calls: " << calls->size() <<endl;
       
       for(auto call_it=calls->begin();call_it!=calls->end();call_it++){
@@ -130,7 +130,7 @@ void SimpleCallCounterProfiler::threadStopped(jobject thread){
 
 void SimpleCallCounterProfiler::reset() {
   for(auto it=statByThread.begin();it!=statByThread.end();it++){
-    unordered_map<unsigned long, CallStatistics*> *calls = it->second;
+    unordered_map<unsigned long long, CallStatistics*> *calls = it->second;
     
     for(auto call_it=calls->begin();call_it!=calls->end();call_it++){
       CallStatistics *stat = call_it->second;
@@ -146,7 +146,7 @@ string SimpleCallCounterProfiler::printCsv(){
   string result = "threadId;methodName;callCount;returnCount\r\n";
   
   for(auto it=statByThread.begin();it!=statByThread.end();it++){
-    unordered_map<unsigned long, CallStatistics*> *calls = it->second;
+    unordered_map<unsigned long long, CallStatistics*> *calls = it->second;
     pthread_t threadId = it->first;
     
     for(auto call_it=calls->begin();call_it!=calls->end();call_it++){
